Fix checkstyle and change UI behaviour
Previously UI got the profile picture from the model. However
now it gets it directly from the data/profilepictures folder using
path in the profilePic field of Person.

This means that the model (ProfilePictureList) is effectively useless
and can be removed. However the command::execute still depends on the
storageManager (uses StorageManager::saveProfilePicture, etc).

Should think of a better way to save the picture that does not break
abstraction