Remove references to JavaFX web
PR #128 included JavaFX web in the `build.gradle` file, which increased
our the filesize of the generated JAR file to close to 100MB. Since we
aren't using any of the features there, the dependency, along with
various references in the test files which weren't being utilized can be
removed. This restores the JAR filesize to ~10MB as well.